+++ src/apps/solar_os_python.c
@@ -3360,7 +3360,7 @@ static mp_obj_t solaros_synth_status(void)
     solar_os_synth_voice_status_t status;
     solar_os_synth_voice_get_status(&status);
 
-    mp_obj_t dict = mp_obj_new_dict(20);
+    mp_obj_t dict = mp_obj_new_dict(27);
     python_dict_store_bool(dict, "running", status.running);
     python_dict_store_cstr(dict, "owner", status.owner);
     python_dict_store_cstr(dict,
@@ -3370,6 +3370,27 @@ static mp_obj_t solaros_synth_status(void)
     python_dict_store_uint(dict, "decay_ms", status.config.decay_ms);
     python_dict_store_uint(dict, "sustain_percent", status.config.sustain_percent);
     python_dict_store_uint(dict, "release_ms", status.config.release_ms);
+    python_dict_store_uint(dict,
+                           "filter_cutoff_hz",
+                           status.config.filter.cutoff_hz);
+    python_dict_store_uint(dict,
+                           "filter_resonance_percent",
+                           status.config.filter.resonance_percent);
+    python_dict_store_uint(dict,
+                           "filter_envelope_amount_percent",
+                           status.config.filter.envelope_amount_percent);
+    python_dict_store_uint(dict,
+                           "filter_attack_ms",
+                           status.config.filter.attack_ms);
+    python_dict_store_uint(dict,
+                           "filter_decay_ms",
+                           status.config.filter.decay_ms);
+    python_dict_store_uint(dict,
+                           "filter_sustain_percent",
+                           status.config.filter.sustain_percent);
+    python_dict_store_uint(dict,
+                           "filter_release_ms",
+                           status.config.filter.release_ms);
     python_dict_store_uint(dict, "active_voices", status.active_voices);
     python_dict_store_uint(dict, "max_voices", SOLAR_OS_SYNTH_VOICE_MAX);
     python_dict_store_uint(dict, "stolen_voices", status.stolen_voices);
@@ -3404,6 +3425,8 @@ static mp_obj_t solaros_synth_configure(size_t n_args, const mp_obj_t *args)
     if (!solar_os_synth_parse_waveform(mp_obj_str_get_str(args[0]), &waveform)) {
         mp_raise_ValueError(MP_ERROR_TEXT("expected square, triangle, saw, or noise"));
     }
+    solar_os_synth_voice_status_t status;
+    solar_os_synth_voice_get_status(&status);
     const solar_os_synth_voice_config_t config = {
         .waveform = waveform,
         .attack_ms = python_optional_u32(n_args,
@@ -3423,6 +3446,7 @@ static mp_obj_t solaros_synth_configure(size_t n_args, const mp_obj_t *args)
                                           args,
                                           4,
                                           SOLAR_OS_SYNTH_VOICE_DEFAULT_RELEASE_MS),
+        .filter = status.config.filter,
     };
     python_check_esp(solar_os_synth_voice_configure(PYTHON_SYNTH_OWNER, &config));
     return mp_const_none;
@@ -3431,6 +3455,58 @@ MP_DEFINE_CONST_FUN_OBJ_VAR_BETWEEN(solaros_synth_configure_obj,
                                     1,
                                     5,
                                     solaros_synth_configure);
+
+static mp_obj_t solaros_synth_configure_filter(size_t n_args,
+                                                const mp_obj_t *args)
+{
+    solar_os_synth_voice_status_t status;
+    solar_os_synth_voice_get_status(&status);
+    solar_os_synth_voice_config_t config = status.config;
+    config.filter = (solar_os_synth_filter_config_t){
+        .cutoff_hz = python_optional_u32(
+            n_args,
+            args,
+            0,
+            SOLAR_OS_SYNTH_VOICE_DEFAULT_FILTER_CUTOFF_HZ),
+        .resonance_percent = python_optional_u8(
+            n_args,
+            args,
+            1,
+            SOLAR_OS_SYNTH_VOICE_DEFAULT_FILTER_RESONANCE_PERCENT),
+        .envelope_amount_percent = python_optional_u8(
+            n_args,
+            args,
+            2,
+            SOLAR_OS_SYNTH_VOICE_DEFAULT_FILTER_ENVELOPE_AMOUNT_PERCENT),
+        .attack_ms = python_optional_u32(
+            n_args,
+            args,
+            3,
+            SOLAR_OS_SYNTH_VOICE_DEFAULT_FILTER_ATTACK_MS),
+        .decay_ms = python_optional_u32(
+            n_args,
+            args,
+            4,
+            SOLAR_OS_SYNTH_VOICE_DEFAULT_FILTER_DECAY_MS),
+        .sustain_percent = python_optional_u8(
+            n_args,
+            args,
+            5,
+            SOLAR_OS_SYNTH_VOICE_DEFAULT_FILTER_SUSTAIN_PERCENT),
+        .release_ms = python_optional_u32(
+            n_args,
+            args,
+            6,
+            SOLAR_OS_SYNTH_VOICE_DEFAULT_FILTER_RELEASE_MS),
+    };
+    python_check_esp(
+        solar_os_synth_voice_configure(PYTHON_SYNTH_OWNER, &config));
+    return mp_const_none;
+}
+MP_DEFINE_CONST_FUN_OBJ_VAR_BETWEEN(solaros_synth_configure_filter_obj,
+                                    1,
+                                    7,
+                                    solaros_synth_configure_filter);
 
 static mp_obj_t solaros_synth_note_on(size_t n_args, const mp_obj_t *args)
 {
@@ -5153,6 +5229,9 @@ static void python_register_solaros_module(void)
     mp_obj_t synth_module = python_new_submodule(module, "synth");
     python_module_store(synth_module, "status", MP_OBJ_FROM_PTR(&solaros_synth_status_obj));
     python_module_store(synth_module, "configure", MP_OBJ_FROM_PTR(&solaros_synth_configure_obj));
+    python_module_store(synth_module,
+                        "configure_filter",
+                        MP_OBJ_FROM_PTR(&solaros_synth_configure_filter_obj));
     python_module_store(synth_module, "note_on", MP_OBJ_FROM_PTR(&solaros_synth_note_on_obj));
     python_module_store(synth_module, "note_off", MP_OBJ_FROM_PTR(&solaros_synth_note_off_obj));
     python_module_store(synth_module,